shuting/[id]/page.tsxのリファクタリング
